Add keys to make map items unique
Before the build [npm run build] was failing due to eslint being active
and some arrays not having keys in React.

Now, the keys were added to the affected arrays. Due to unfamiliarity
with the objects at hand, a simple id of `itemName{index}` was used.

These changes should allow Vercel to finish the build process and deploy
without issues.

It is important to have keys in React map items because keys help React
identify which items have changed, are added, or are removed. Keys
should be given to the elements inside the array to give the elements a
stable identity. This improves the performance and efficiency of
rendering lists of elements in React.

https://react.dev/learn/rendering-lists