Add content-based filtering recommendation system
The system suggests songs by calculating the Euclidean distance between
the mean musical characteristics of each MBTI personality type and
individual tracks in the two datasetks.

In simpler terms, if user A belongs to type X, and type X
fancies tracks with high energy, they will be suggested tracks with high energy.

The algorithm doesn't yet consider the user's liked songs. Once we've
successfully integrated the ML environment with the application, I'll
begin exploring the inclusion of the user's liked songs in the
recommendations.