Use state to manage liked songs locally
In addition to updating the DB whenever a song is added or removed, a
state was also added so that this change is reflected immediately rather
than relying solely on the DB updates.

Before, the user had to wait for the DB query to complete for the
changes to be reflected on the page.

Now, the addition and removal of the liked songs is handled by local
state.